Import 1.3.26

+++ include/net/ipip.h
@@ -1,4 +1,4 @@
 extern int ipip_rcv(struct sk_buff *skb, struct device *dev, struct options *opt, 
-		unsigned long daddr, unsigned short len, unsigned long saddr,
+		__u32 daddr, unsigned short len, __u32 saddr,
                                    int redo, struct inet_protocol *protocol);
